Add a namespace and folder detail (issue #17)
